add i32

+++ core/src/data/value/binary_op/mod.rs
@@ -2,10 +2,10 @@ use crate::{prelude::Value, result::Result};
 
 mod decimal;
 mod f64;
-mod i8;
+mod i128;
 mod i32;
 mod i64;
-mod i128;
+mod i8;
 
 pub trait TryBinaryOperator {
     type Rhs;
